Fix drain_all infinite loop after clearing worker with symbolized queue
`jobs_by_queue` saves the queue name of enqueued jobs as strings. When
`clear` is called for a worker that has it's queue name defined as a
symbol `clear` calls `Queues.clear_for` using a symbol. As a result
`jobs_by_queue` is not cleared, while `jobs_by_class` is. Calling
`drain_all` then runs into an infinite loop because of the mismatch of
the two state hashes.